refactor(sse): extract Kiro tool_call validation and stream error formatting to leaf modules
open-sse/executors/kiro.ts (1052 lines) landed above the 1000-line new-file file-size cap
and open-sse/utils/stream.ts (2941 lines) grew past its frozen 2889-line baseline, both
from #9314's nested tool_call validation fix. Decompose instead of rebaselining:

- Move the pure Kiro tool_call wrapper validation/buffering helpers
  (validateKiroToolUse, validateKiroToolCallWrapperInput, appendBufferedKiroToolInput,
  etc.) to open-sse/executors/kiroToolCallValidation.ts. kiro.ts re-exports
  validateKiroToolUse to keep its public surface unchanged.
- Move the upstream stream-failure normalization + client-format error framing
  (normalizeStreamFailurePayload, formatTranslatedStreamError) to
  open-sse/utils/streamErrorFormat.ts.

Behavior is unchanged — same functions, same logic, just relocated. Verified via
tests/unit/kiro-tool-call-validation.test.ts and the stream.ts test suite.
